requirements.txt
fastapi
uvicorn[standard]
python-dotenv
prometheus-client

sqlalchemy
psycopg2-binary

langchain
langchain-openai
langgraph
tenacity
chromadb
pydantic
email-validator

